+++ noodles-cram/CHANGELOG.md
@@ -7,6 +7,9 @@
   * cram/reader/data_container/compression_header/encoding: Add Golomb and
     Golomb-Rice encoding decoders.
 
+  * cram/writer/data_container/compression_header/encoding: Add Golomb and
+    Golomb-Rice encoding writers.
+
     These see no usage in practice, but they are still in the spec (§ 13.8 and
     § 13.9, respectively). They are only added for completeness.